% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_has_access}
\alias{check_has_access}
\title{Check if user is in hscdiip}
\usage{
check_has_access(group)
}
\arguments{
\item{group}{The group to check}
}
\value{
Boolean
}
\description{
Check if user is in hscdiip
}
