% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demog_vars}
\alias{demog_vars}
\title{Demographic variables}
\format{
An object of class \code{character} of length 30.
}
\usage{
demog_vars
}
\description{
A vector containing the names of the demographic variables.
}
\keyword{data}
