% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ep_file_bedday_vars}
\alias{ep_file_bedday_vars}
\title{Episode file bedday variables}
\format{
An object of class \code{character} of length 14.
}
\usage{
ep_file_bedday_vars
}
\description{
A vector containing the names of the
bedday related variables in the
episode file.
}
\keyword{data}
