% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf.R
\name{format_year}
\alias{format_year}
\title{Title}
\usage{
format_year(year)
}
\arguments{
\item{year}{Year in any format}
}
\value{
Year correctly formatted as "CCYY"
}
\description{
Title
}
\examples{
format_year("2017/18")
}
