% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anon_chi.R
\name{get_anon_chi}
\alias{get_anon_chi}
\title{Match on anon_chi to an object using CHI numbers}
\usage{
get_anon_chi(data, chi_var = "chi", drop = TRUE)
}
\arguments{
\item{data}{tibble or data frame}

\item{chi_var}{CHI variable: the name of the variable containg CHI (default is chi)}

\item{drop}{Optional boolean indicating whether the existing chi_var should be dropped - default is TRUE}
}
\value{
a tibble
}
\description{
Match on anon_chi to an object using CHI numbers
}
\examples{
get_anon_chi(data)
get_anon_chi(data, chi_var = "uniqueid")
}
