% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chi.R
\name{get_chi}
\alias{get_chi}
\title{Match on CHI numbers to a dataset using anon_chi}
\usage{
get_chi(data, anon_chi_var = "anon_chi", drop = TRUE)
}
\arguments{
\item{data}{tibble or data frame}

\item{anon_chi_var}{Anon CHI variable: the name of the variable
containing anonymised CHI (default is anon_chi)}

\item{drop}{Optional boolean indicating whether the existing \code{anon_chi_var}
should be dropped - default is TRUE}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Match on CHI numbers to a dataset using anon_chi
}
\examples{
\dontrun{
slf_1718 <- read_slf_individual("1718", from = 1, to = 100)
get_chi(slf_1718)
get_chi(slf_1718, drop = FALSE)
}
}
