% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ltc_vars}
\alias{ltc_vars}
\title{LTC variables}
\format{
An object of class \code{character} of length 19.
}
\usage{
ltc_vars
}
\description{
A vector containing the names of the
Long Term Condition (LTC) variables.
}
\keyword{data}
