% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf.R
\name{read__slf_episode}
\alias{read__slf_episode}
\title{Read a Source Linkage episode file}
\usage{
read__slf_episode(year, ...)
}
\arguments{
\item{year}{of the file you would like to load}

\item{...}{other options to be passed to read_fst()}
}
\value{
a tibble
}
\description{
Read a Source Linkage episode file
}
\examples{
read__slf_episode("1718")
read__slf_episode("1718", columns = c("anon_chi", "dob", "demographic_cohort"), from = 100000, to = 200000)
}
