% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf.R
\name{read__slf_individual}
\alias{read__slf_individual}
\title{Read a Source Linkage individual file}
\usage{
read__slf_individual(year, ...)
}
\arguments{
\item{year}{of the file you would like to load}

\item{...}{other options to be passed to read_fst()}
}
\value{
a tibble
}
\description{
Read a Source Linkage individual file
}
\examples{
read_slf_individual("1718")
read_slf_individual("1718", columns = c("anon_chi", "dob", "hri_scot"), from = 100000, to = 200000))
}
