% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf.R
\name{read_slf}
\alias{read_slf}
\title{Read a Source Linkage File}
\usage{
read_slf(year, file_version, ...)
}
\arguments{
\item{year}{Year of the file to be read}

\item{file_version}{Version of the file (Individual / episode)}

\item{...}{other options to be passed to read_fst()}
}
\value{
a tibble
}
\description{
Read a Source Linkage File
}
\examples{
read_slf("1718", "episode")
}
