% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf_parquet.R
\name{read_slf_episode_parquet}
\alias{read_slf_episode_parquet}
\title{Read a Source Linkage episode file (parquet)}
\usage{
read_slf_episode_parquet(
  year,
  col_select = NULL,
  partnerships = NULL,
  recids = NULL,
  as_data_frame = TRUE,
  dev = TRUE,
  columns = lifecycle::deprecated()
)
}
\arguments{
\item{year}{Year of the file to be read, you can specify multiple years
which will then be returned as one file. It will be converted to short FY
using \code{\link[=format_year]{format_year()}}.}

\item{col_select}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:eval_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}

\item{partnerships}{Optional specify a partnership (hscp2018) or
partnerships to select}

\item{recids}{Optional specify a recid or recids to select}

\item{as_data_frame}{Should the function return a \code{data.frame} (default) or
an Arrow \link[arrow]{Table}?}

\item{dev}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Whether to get the file from
the development area (\verb{/conf/sourcedev/Source_Linkage_File_Updates}).
The default (\code{FALSE}) will get the production file from the usual area.}

\item{columns}{Column names to read. The default is to read all columns.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
read_slf_episode("1718",
  columns = c("anon_chi", "dob", "demographic_cohort"),
  from = 100000,
  to = 101000
)

read_slf_episode(c("1718", "1819"),
  columns = c("anon_chi", "dob", "demographic_cohort"),
  from = 100000,
  to = 101000
)
}
}
