% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf.R
\name{read_slf_individual}
\alias{read_slf_individual}
\title{Read a Source Linkage individual file}
\usage{
read_slf_individual(year, columns = NULL, partnership = NA, ...)
}
\arguments{
\item{year}{Year of the file to be readl you can specify multiple years
which will then be returned as one file}

\item{columns}{Optional a vector of column names}

\item{partnership}{Optional specify Partnership to select}

\item{...}{other options to be passed to read_fst()}
}
\value{
a tibble
}
\description{
Read a Source Linkage individual file
}
\examples{
read_slf_individual("1718")
read_slf_individual("1718",
  columns = c("anon_chi", "dob", "hri_scot"),
  from = 100000, to = 200000
)
read_slf_individual(c("1516", "1617", "1718", "1819"),
  columns = c("anon_chi", "dob", "hri_scot")
)
}
