% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf_parquet.R
\name{read_slf_parquet}
\alias{read_slf_parquet}
\title{Read a Source Linkage File (parquet)}
\usage{
read_slf_parquet(
  year,
  col_select = NULL,
  as_data_frame = TRUE,
  file_version = c("episode", "individual"),
  partnerships = NULL,
  recids = NULL,
  dev = TRUE
)
}
\arguments{
\item{year}{Year of the file to be read, you can specify multiple years
which will then be returned as one file. It will be converted to short FY
using \code{\link[=format_year]{format_year()}}.}

\item{col_select}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:eval_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}

\item{as_data_frame}{Should the function return a \code{data.frame} (default) or
an Arrow \link[arrow]{Table}?}

\item{file_version}{Version of the file (individual / episode)}

\item{partnerships}{Optional specify a partnership (hscp2018) or
partnerships to select}

\item{recids}{Optional specify a recid or recids to select}

\item{dev}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Whether to get the file from
the development area (\verb{/conf/sourcedev/Source_Linkage_File_Updates}).
The default (\code{FALSE}) will get the production file from the usual area.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
