% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recids}
\alias{recids}
\title{Recid name / description lookup}
\format{
A \link[tibble:tibble-package]{tibble} with 18 rows and 2 variables:
\describe{
\item{recid}{The short-form recid}
\item{Description}{A description of the recid}
}
}
\usage{
recids
}
\description{
A \link[tibble:tibble-package]{tibble} containing the names and description of the recids.
To be used with the \code{recid} parameter when reading the episode files.
}
\keyword{data}
