% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_has_access}
\alias{check_has_access}
\title{Check if the current user is in a UNIX group}
\usage{
check_has_access(group = "hscdiip")
}
\arguments{
\item{group}{The group to check, default is hscdiip}
}
\value{
Boolean
}
\description{
Check if the current user is in a UNIX group
}
