% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_year.R
\name{check_year}
\alias{check_year}
\title{Check the year(s) supplied are valid}
\usage{
check_year(year)
}
\arguments{
\item{year}{Year of the file to be read, you can specify multiple years
which will then be returned as one file. It will be converted to short FY
using \code{\link[=format_year]{format_year()}}.}
}
\value{
\code{TRUE} if the year is valid and a newer year, \code{FALSE} for any
older but still valid years. Otherwise throws an error.
}
\description{
Check the year(s) supplied are valid
}
