% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_year.R
\name{check_year}
\alias{check_year}
\title{Check the year(s) supplied are valid}
\usage{
check_year(year)
}
\arguments{
\item{year}{A year in the short FY format use \code{format_year()}}
}
\value{
TRUE if pass otherwise an error message
}
\description{
Check the year(s) supplied are valid
}
