% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ep_file_vars}
\alias{ep_file_vars}
\title{Names of variables in the epsiode files}
\format{An object of class \code{character} of length 210.}
\usage{
ep_file_vars
}
\description{
Names of variables in the epsiode files
}
\keyword{data}
