% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ep_file_vars}
\alias{ep_file_vars}
\title{Episode file variables}
\format{
An object of class \code{list} of length 238.
}
\usage{
ep_file_vars
}
\description{
The names of the variables in the episode files,
in the correct order.
}
\keyword{data}
