% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_year.R
\name{format_year}
\alias{format_year}
\title{Format a year into the correct short FY format}
\usage{
format_year(year)
}
\arguments{
\item{year}{Year in any format}
}
\value{
Year correctly formatted as short FY
}
\description{
Format a year into the correct short FY format
}
