% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_year.R
\name{format_year}
\alias{format_year}
\title{Format Year parameter}
\usage{
format_year(year)
}
\arguments{
\item{year}{Year in any format}
}
\value{
Year correctly formatted as "CCYY"
}
\description{
Format Year parameter
}
