% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_year.R
\name{format_year}
\alias{format_year}
\title{Format a year into the correct short FY format}
\usage{
format_year(year)
}
\arguments{
\item{year}{Year of the file to be read, you can specify multiple years
which will then be returned as one file. It will be converted to short FY
using \code{\link[=format_year]{format_year()}}.}
}
\value{
Year correctly formatted into short FY e.g. '1718'
}
\description{
Format a year into the correct short FY format
}
