% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_file_path.R
\name{gen_file_path}
\alias{gen_file_path}
\title{Generate a Source Linkage File, file path}
\usage{
gen_file_path(year, file_version)
}
\arguments{
\item{year}{Year of the file to be read, you can specify multiple years
which will then be returned as one file. It will be converted to short FY
using \code{\link[=format_year]{format_year()}}.}

\item{file_version}{Version of the file (individual / episode)}
}
\value{
a \link[fs:fs-package]{fs_path} object
}
\description{
Generate a Source Linkage File, file path
}
