% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anon_chi.R
\name{get_anon_chi}
\alias{get_anon_chi}
\title{Match on anon_chi to a dataset using CHI numbers}
\usage{
get_anon_chi(tibb, chi_var = "chi", drop = TRUE)
}
\arguments{
\item{tibb}{tibble or data frame}

\item{chi_var}{CHI variable: the name of the variable containing CHI
(default is chi)}

\item{drop}{Optional boolean indicating whether the existing chi_var
should be dropped - default is TRUE}
}
\value{
a tibble
}
\description{
Match on anon_chi to a dataset using CHI numbers
}
\examples{
chi_cohort \%>\% get_anon_chi()
chi_cohort \%>\% get_anon_chi(chi_var = "upi_number")
}
