% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chi.R
\name{get_chi}
\alias{get_chi}
\title{Match on CHI numbers to a dataset using anon_chi}
\usage{
get_chi(data, anon_chi_var = "anon_chi", drop = TRUE)
}
\arguments{
\item{data}{tibble or data frame}

\item{anon_chi_var}{Anon CHI variable: the name of the variable containg anonymised CHI (default is anon_chi)}

\item{drop}{Optional boolean indicating whether the existing anon_ch_var should be dropped - default is TRUE}
}
\value{
a tibble
}
\description{
Match on CHI numbers to a dataset using anon_chi
}
\examples{
get_chi(data)
get_chi(data, drop = FALSE)
}
