% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiv_file_vars}
\alias{indiv_file_vars}
\title{Individual file variables}
\format{
An object of class \code{character} of length 205.
}
\usage{
indiv_file_vars
}
\description{
The names of the variables in the individual files,
in the correct order.
}
\keyword{data}
