% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiv_file_vars}
\alias{indiv_file_vars}
\title{Names of variables in the individual files}
\format{
An object of class \code{character} of length 180.
}
\usage{
indiv_file_vars
}
\description{
Names of variables in the individual files
}
\keyword{data}
