% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partnerships}
\alias{partnerships}
\title{HSCP name / code lookup}
\format{
A \link[tibble:tibble-package]{tibble} with 31 rows and 2 variables:
\describe{
\item{partnership_name}{The official name of the Partnership}
\item{hscp2018}{The hscp2018 standard code S370000??}
}
}
\usage{
partnerships
}
\description{
A \link[tibble:tibble-package]{tibble}
containing the names and codes (hscp2018 configuration) of partnerships.
To be used with the \code{partnership} parameter when reading files.
}
\keyword{data}
