% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partnerships}
\alias{partnerships}
\title{A tibble containing the names and codes (2018 configuration) of parterships
to be used with the partnership parameter when reading files}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 2 columns.
}
\usage{
partnerships
}
\description{
A tibble containing the names and codes (2018 configuration) of parterships
to be used with the partnership parameter when reading files
}
\keyword{data}
