% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf.R
\name{read_slf}
\alias{read_slf}
\title{Read a Source Linkage File}
\usage{
read_slf(
  year,
  file_version = c("episode", "individual"),
  partnerships = NULL,
  recids = NULL,
  ...
)
}
\arguments{
\item{year}{Year of the file to be read, you can specify multiple years
which will then be returned as one file. It will be converted to short FY
using \code{\link[=format_year]{format_year()}}.}

\item{file_version}{Version of the file (individual / episode)}

\item{partnerships}{Optional specify a partnership (hscp2018) or
partnerships to select}

\item{recids}{Optional specify a recid or recids to select}

\item{...}{other options to be passed to \code{\link[fst:write_fst]{fst::read_fst()}}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Read a Source Linkage File
}
