% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf.R
\name{read_slf_episode}
\alias{read_slf_episode}
\title{Read a Source Linkage episode file}
\usage{
read_slf_episode(year, columns = NULL, partnerships = NULL, recids = NULL, ...)
}
\arguments{
\item{year}{Year of the file to be read you can specify multiple years
which will then be returned as one file}

\item{columns}{Optional a vector of column names}

\item{partnerships}{Optional specify a partnership (hscp2018) or
partnerships to select}

\item{recids}{Optional specify a recid or recids to select}

\item{...}{other options to be passed to read_fst()}
}
\value{
a tibble
}
\description{
Read a Source Linkage episode file
}
\examples{
read_slf_episode("1718",
  columns = c("anon_chi", "dob", "demographic_cohort"),
  from = 100000,
  to = 101000
)

read_slf_episode(c("1718", "1819"),
  columns = c("anon_chi", "dob", "demographic_cohort"),
  from = 100000,
  to = 101000
)
}
