% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf.R
\name{read_slf_individual}
\alias{read_slf_individual}
\title{Read a Source Linkage individual file}
\usage{
read_slf_individual(year, columns = NULL, partnerships = NULL, ...)
}
\arguments{
\item{year}{Year of the file to be read, you can specify multiple years
which will then be returned as one file. It will be converted to short FY
using \code{\link[=format_year]{format_year()}}.}

\item{columns}{Column names to read. The default is to read all columns.}

\item{partnerships}{Optional specify a partnership (hscp2018) or
partnerships to select}

\item{...}{other options to be passed to \code{\link[fst:write_fst]{fst::read_fst()}}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Read a Source Linkage individual file
}
\examples{
\dontrun{
read_slf_individual("1718",
  columns = c("anon_chi", "dob", "hri_scot"),
  from = 100000, to = 101000
)
read_slf_individual(c("1718", "1819"),
  columns = c("anon_chi", "dob", "hri_scot"),
  from = 100000,
  to = 101000
)
}
}
