% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_partnerships}
\alias{sc_partnerships}
\title{Social Care Partnership names / sending location code / LCA code lookup}
\format{
A \link[tibble:tibble-package]{tibble} with 32 rows and 3 variables:
\describe{
\item{partnership_name}{The official name of the Partnership}
\item{lca}{The coresponding LCA code 05L}
\item{sending_location}{The sending_location standard code 100L}
}
}
\usage{
sc_partnerships
}
\description{
A \link[tibble:tibble-package]{tibble}
containing the names and codes (sending location/LCA configuration) of social care
partnerships.
}
\details{
To be used with the \code{sc_partnerships} parameter when reading files.
}
\keyword{data}
