% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slfhelper-package.R
\docType{package}
\name{slfhelper-package}
\alias{slfhelper}
\alias{slfhelper-package}
\title{slfhelper: Useful functions for working with the Source Linkage Files}
\description{
This package provides helper functions for working with the Source Linkage Files (SLFs). The functions are mainly focused on making the first steps of analysis easier. They can read and filter the files efficiently using minimal code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://public-health-scotland.github.io/slfhelper/}
  \item \url{https://github.com/Public-Health-Scotland/slfhelper}
  \item Report bugs at \url{https://github.com/Public-Health-Scotland/slfhelper/issues}
}

}
\author{
\strong{Maintainer}: Megan McNicol \email{megan.mcnicol2@phs.scot}

Authors:
\itemize{
  \item James McMahon \email{james.mcmahon@phs.scot} (\href{https://orcid.org/0000-0002-5380-2029}{ORCID})
  \item Zihao Li \email{zihao.li@phs.scot} (\href{https://orcid.org/0000-0002-5178-2124}{ORCID})
  \item Jennifer Thom \email{jennifer.thom@phs.scot}
}

Other contributors:
\itemize{
  \item Public Health Scotland \email{phs.source@phs.scot} [copyright holder]
}

}
\keyword{internal}
