% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slfhelper-package.R
\docType{package}
\name{slfhelper-package}
\alias{slfhelper}
\alias{slfhelper-package}
\title{slfhelper: Useful functions for working with the Source Linkage Files}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Moohan/slfhelper}
  \item Report bugs at \url{https://github.com/Moohan/slfhelper/issues}
}

}
\keyword{internal}
